Be more careful about the semantics of `count` and wildcard columns (#558)
Previously, rows that only differed on wildcards would not be counted correctly (they would be deduplicated in join because join currently uses set semantics). This also changes the syntax for `count` to clarify that we are not counting the unique number of anything, but instead are counting how many rows are in the group.